Inventory Project Management modifications
- Added ability to delete an inventory project
- Bug fix: issue interferred with saving changes to an existing project
- Misc modifications that improve management workflow

+++ classes/InventoryProjectManager.php
@@ -10,8 +10,8 @@ class InventoryProjectManager {
 	private $isPublic = 1;
 	private $errorStr;
 
-	public function __construct(){
-		$this->conn = MySQLiConnectionFactory::getCon("readonly");
+	public function __construct($connType = 'readonly'){
+		$this->conn = MySQLiConnectionFactory::getCon($connType);
 		$this->googleUrl = "http://maps.google.com/maps/api/staticmap?size=120x150&maptype=terrain";
 		if(array_key_exists('GOOGLE_MAP_KEY',$GLOBALS) && $GLOBALS['GOOGLE_MAP_KEY']) $this->googleUrl .= '&key='.$GLOBALS['GOOGLE_MAP_KEY'];
 	}
@@ -44,7 +44,6 @@ class InventoryProjectManager {
 				'occurrencesearch, ispublic, sortsequence '.
 				'FROM fmprojects '.
 				'WHERE (pid = '.$this->pid.') ';
-			//echo $sql;
 			$rs = $this->conn->query($sql);
 			if($row = $rs->fetch_object()){
 				$this->pid = $row->pid;
@@ -65,8 +64,7 @@ class InventoryProjectManager {
 	}
 
 	public function submitProjEdits($projArr){
-		$conn = MySQLiConnectionFactory::getCon("write");
-		$fieldArr = array('projname', 'displayname', 'managers', 'fulldescription', 'notes', 'ispublic', 'parentpid', 'SortSequence');
+		$fieldArr = array('projname', 'displayname', 'managers', 'fulldescription', 'notes', 'ispublic', 'parentpid', 'sortsequence');
 		$sql = "";
 		foreach($projArr as $field => $value){
 			if(in_array($field,$fieldArr)){
@@ -75,51 +73,66 @@ class InventoryProjectManager {
 		}
 		$sql = 'UPDATE fmprojects SET '.substr($sql,1).' WHERE (pid = '.$this->pid.')';
 		//echo $sql; exit;
-		$conn->query($sql);
-		$conn->close();
+		$this->conn->query($sql);
 	}
-	
-	public function addNewProject($projArr){
-		$conn = MySQLiConnectionFactory::getCon("write");
-		$sql = 'INSERT INTO fmprojects(projname,managers,fulldescription,notes,ispublic,sortsequence) '.
-			'VALUES("'.$this->cleanInStr($projArr['projname']).'","'.$this->cleanInStr($projArr['managers']).'","'.
-			$this->cleanInStr($projArr['fulldescription']).'","'.
-			$this->cleanInStr($projArr['notes']).'",'.$this->cleanInStr($projArr['ispublic']).','.
-			(is_numeric($projArr['sortsequence'])?$projArr['sortsequence']:'50').')';
-		//echo $sql;
-		if($conn->query($sql)){
-			$this->pid = $conn->insert_id;
-			
+
+	public function deleteProject($projID){
+		$status = true;
+		if($projID && is_numeric($projID)){
+			$sql = 'DELETE FROM fmprojects WHERE pid = '.$projID;
+			//echo $sql; exit;
+			if(!$this->conn->query($sql)){
+				$status = false;
+				$this->errorStr = 'ERROR deleting inventory project: '.$this->conn->error;
+			}
 		}
-		$conn->close();
+		return $status;
+	}
+
+	public function addNewProject($projArr){
+		$sql = 'INSERT INTO fmprojects(projname,managers,fulldescription,notes,ispublic) '.
+			'VALUES("'.$this->cleanInStr($projArr['projname']).'",'.
+			($projArr['managers']?'"'.$this->cleanInStr($projArr['managers']).'"':'NULL').','.
+			($projArr['fulldescription']?'"'.$this->cleanInStr($projArr['fulldescription']).'"':'NULL').','.
+			($projArr['notes']?'"'.$this->cleanInStr($projArr['notes']).'"':'NULL').','.
+			(is_numeric($projArr['ispublic'])?$projArr['ispublic']:'0').')';
+		//echo $sql;
+		if($this->conn->query($sql)){
+			$this->pid = $this->conn->insert_id;
+		}
+		else{
+			$this->errorStr = 'ERROR creating new project: '.$this->conn->error;
+		}
 		return $this->pid;
 	}
 	
 	public function getResearchChecklists(){
 		global $USER_RIGHTS;
-		$returnArr = Array();
-		$sql = 'SELECT c.clid, c.name, c.latcentroid, c.longcentroid, c.access '.
-			'FROM fmchklstprojlink cpl INNER JOIN fmchecklists c ON cpl.clid = c.clid '.
-			'WHERE (cpl.pid = '.$this->pid.') AND ((c.access != "private")';
-		if(array_key_exists('ClAdmin',$USER_RIGHTS)){
-			$sql .= ' OR (c.clid IN ('.implode(',',$USER_RIGHTS['ClAdmin']).'))) ';
-		}
-		else{
-			$sql .= ') ';
-		}
-		$sql .= "ORDER BY c.SortSequence, c.name";
-		//echo $sql;
-		$rs = $this->conn->query($sql);
-		$cnt = 0;
-		while($row = $rs->fetch_object()){
-			$returnArr[$row->clid] = $this->cleanOutStr($row->name).($row->access == 'private'?' <span title="Viewable only to editors">(private)</span>':'');
-			if($cnt < 50 && $row->latcentroid){
-				$this->researchCoord[] = $row->latcentroid.','.$row->longcentroid;
+		$retArr = Array();
+		if($this->pid){
+			$sql = 'SELECT c.clid, c.name, c.latcentroid, c.longcentroid, c.access '.
+				'FROM fmchklstprojlink cpl INNER JOIN fmchecklists c ON cpl.clid = c.clid '.
+				'WHERE (cpl.pid = '.$this->pid.') AND ((c.access != "private")';
+			if(array_key_exists('ClAdmin',$USER_RIGHTS)){
+				$sql .= ' OR (c.clid IN ('.implode(',',$USER_RIGHTS['ClAdmin']).'))) ';
 			}
-			$cnt++;
+			else{
+				$sql .= ') ';
+			}
+			$sql .= "ORDER BY c.SortSequence, c.name";
+			//echo $sql;
+			$rs = $this->conn->query($sql);
+			$cnt = 0;
+			while($row = $rs->fetch_object()){
+				$retArr[$row->clid] = $this->cleanOutStr($row->name).($row->access == 'private'?' <span title="Viewable only to editors">(private)</span>':'');
+				if($cnt < 50 && $row->latcentroid){
+					$this->researchCoord[] = $row->latcentroid.','.$row->longcentroid;
+				}
+				$cnt++;
+			}
+			$rs->free();
 		}
-		$rs->free();
-		return $returnArr;
+		return $retArr;
 	}
 	
 	public function getGoogleStaticMap(){
@@ -134,49 +147,41 @@ class InventoryProjectManager {
 	//User management functions
 	public function getManagers(){
 		$retArr = array();
-		$sql = 'SELECT u.uid, CONCAT_WS(", ", u.lastname, u.firstname) as fullname, l.username '.
-			'FROM userroles r INNER JOIN users u ON r.uid = u.uid '.
-			'INNER JOIN userlogin l ON u.uid = l.uid '.
-			'WHERE r.role = "ProjAdmin" AND r.tablepk = '.$this->pid;
-		$rs = $this->conn->query($sql);
-		while($r = $rs->fetch_object()){
-			$retArr[$r->uid] = $r->fullname.' ('.$r->username.')';
+		if($this->pid){
+			$sql = 'SELECT u.uid, CONCAT_WS(", ", u.lastname, u.firstname) as fullname, l.username '.
+				'FROM userroles r INNER JOIN users u ON r.uid = u.uid '.
+				'INNER JOIN userlogin l ON u.uid = l.uid '.
+				'WHERE r.role = "ProjAdmin" AND r.tablepk = '.$this->pid;
+			$rs = $this->conn->query($sql);
+			while($r = $rs->fetch_object()){
+				$retArr[$r->uid] = $r->fullname.' ('.$r->username.')';
+			}
+			$rs->free();
+			asort($retArr);
 		}
-		$rs->free();
-		asort($retArr);
 		return $retArr;
 	} 
 	
 	public function addManager($uid){
 		$status = false;
 		if(is_numeric($uid) && $this->pid){
-			$conn = MySQLiConnectionFactory::getCon("write");
 			$sql = 'INSERT INTO userroles(role,tablename,tablepk,uid) '.
 				'VALUES("ProjAdmin","fmprojects",'.$this->pid.','.$uid.') ';
-			if($conn->query($sql)){
-				
+			if(!$this->conn->query($sql)){
+				$this->errorStr = 'ERROR adding manager: '.$this->conn->error;
 			}
-			else{
-				$this->errorStr = 'ERROR adding manager: '.$conn->error;
-			}
-			if(!($conn === null)) $conn->close();
 		}
 		return $status;
 	} 
 	
 	public function deleteManager($uid){
-		$status = false;
+		$status = true;
 		if(is_numeric($uid) && $this->pid){
-			$conn = MySQLiConnectionFactory::getCon("write");
-			$sql = 'DELETE FROM userroles '.
-				'WHERE (role = "ProjAdmin") AND (tablepk = '.$this->pid.') AND (uid = '.$uid.') ';
-			if($conn->query($sql)){
-				$status = true;
+			$sql = 'DELETE FROM userroles WHERE (role = "ProjAdmin") AND (tablepk = '.$this->pid.') AND (uid = '.$uid.') ';
+			//echo $sql;
+			if(!$this->conn->query($sql)){
+				$this->errorStr = 'ERROR removing manager: '.$this->conn->error;
 			}
-			else{
-				$this->errorStr = 'ERROR removing manager: '.$conn->error;
-			}
-			if(!($conn === null)) $conn->close();
 		}
 		return $status;
 	}
@@ -198,28 +203,18 @@ class InventoryProjectManager {
 	//Checklist management functions
 	public function addChecklist($clid){
 		if(!is_numeric($clid)) return false; 
-		$conn = MySQLiConnectionFactory::getCon("write");
 		$sql = 'INSERT INTO fmchklstprojlink(pid,clid) VALUES('.$this->pid.','.$clid.') ';
-		if($conn->query($sql)){
-			return 'SUCCESS: Checklist has been added to project';
+		if(!$this->conn->query($sql)){
+			return 'ERROR adding checklist to project: '.$this->conn->error;
 		}
-		else{
-			return 'FAILED: Unable to add checklist to project';
-		}
-		if(!($conn === null)) $conn->close();
 	}
 
 	public function deleteChecklist($clid){
 		if(!is_numeric($clid)) return false; 
-		$conn = MySQLiConnectionFactory::getCon("write");
 		$sql = 'DELETE FROM fmchklstprojlink WHERE (pid = '.$this->pid.') AND (clid = '.$clid.')';
-		if($conn->query($sql)){
-			return 'SUCCESS: Checklist has been deleted from project';
+		if($this->conn->query($sql)){
+			return 'ERROR deleting checklist from project';
 		}
-		else{
-			return 'FAILED: Unable to checklist from project';
-		}
-		if(!($conn === null)) $conn->close();
 	}
 
 	//Misc functions
